hyper-v: fix bad type
(cherry picked from commit 9baa3df90a6716b47b714ff79f929599df23a96c)

+++ core/src/win32/plugins/filed/hyper-v.cc
@@ -1777,7 +1777,7 @@ class VirtualSystemSnapshotService : ClassObject {
   static constexpr std::wstring_view class_name{
       L"Msvm_VirtualSystemSnapshotService"};
 
-  enum class SnapshotType : std::int32_t
+  enum class SnapshotType : std::uint16_t
   {
     FullSnapshot = 2,
     DiskSnapshot = 3,
